setup -dev application properties file and configured AWS RDS

+++ src/main/java/com/davault/fullstack/student/StudentService.java
@@ -1,7 +1,8 @@
 package com.davault.fullstack.student;
 
+import com.davault.fullstack.student.exception.BadRequestException;
+import com.davault.fullstack.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,12 +18,21 @@ public class StudentService {
     }
 
     public void addStudent(Student student) {
-        // check if email is taken
+        Boolean existsEmail = studentRepository
+                .selectExistsEmail(student.getEmail());
+        if (existsEmail) {
+            throw new BadRequestException(
+                    "Email " + student.getEmail() + " taken");
+        }
+
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
-        // check if student exists
+        if(!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Student with id " + studentId + " does not exists");
+        }
         studentRepository.deleteById(studentId);
     }
 }
